Algorithm study day 18
- 완전 탐색 알고리즘 학습
- 프로젝트 구조 변경
- 백준 15649 N과 M(1) : https://www.acmicpc.net/problem/15649
- 백준 15650 N과 M (2) : https://www.acmicpc.net/problem/15650
- 백준 15651 N과 M (3) : https://www.acmicpc.net/problem/15651
- 백준 15652 N과 M (4) : https://www.acmicpc.net/problem/15652
